Add preloadable sites not yet preloaded in Tor Browser
This adds a large collection of rulesets for hosts that are either
in the Chromium preload list [1], or are pending inclusion in that list [2],
but which are not yet in the Tor Browser preload list [3]. Concensus
these days appears to be that rulesets should be deleted whenever
their hosts end up on the Tor preload list, but that until that
happens, it's fine to have them in HTTPS Everywhere. I personally
agree with that as it does make using HTTP Nowhere more pleasant.

Indeed, it happens every now and then that someone submits a PR
for a preloadable-but-not-preloaded site, so the main point of this
commit is to enforce consistency and to take care of those once and
for all.

On the other hand, some amount of care has gone into making sure
that this stays on the conservative side. In particular,
* this includes only eTLD+1s,
* this includes only hosts that specify includeSubdomains,
* this excludes all domains which are not already in the HTTPS-E db
  in one way or another (even if they are not active), and
* this does not introduce wildcard rules, even though it probably
  could: The main reason for removal from the preload list is subdomain
  breakage.

[1]: https://chromium.googlesource.com/chromium/src/+/master/net/http/transport_security_state_static.json
[2]: https://hstspreload.appspot.com/pending
[3]: https://gitweb.torproject.org/tor-browser.git/plain/security/manager/boot/src/nsSTSPreloadList.inc?h=tor-browser-38.6.1esr-6.0-1